% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_functions.R
\name{classify_sentiment}
\alias{classify_sentiment}
\title{Classify Sentiment}
\usage{
classify_sentiment(score)
}
\arguments{
\item{score}{A numeric value of sentiment score.}
}
\description{
This function classifies the sentiment as positive (1) or negative (0) based on sentiment score.
}
\examples{
classify_sentiment(3) # Returns 1
classify_sentiment(-1) # Returns 0
}
